Added Class Requirements Functionality
1. Mouse support added but buggy
2. Controller Bug Fixed
3. Start and Quit functionality added along with start scene
4. Added a line that gets bigger as camera starts moving faster.

Other:
Changed fire button to alter color of cube material.

Made a lot of code into functions and seperated out dependancies within the class to make adjustments easier to make and sequencing less important.

Changed the deadzone to be dependent on input rather than angle. Change makes finding deadzone start point easier.

upon camera unlocking will now start slowly accelerating instead of instantly starting at the speed it'd be at if there was no dead zone.

Should be configured to make a working build with start and exit functionality as well as automatic transitions between mouse and controller if controller is turned off or on.

Removed Player movement code as it was not immediately part of the project.